Some Code Arrangeing
